refactor(core): decouple GenerationResult from runtimes base
